TST: reorg of DatetimeIndex tests from test_timeseries.py to test_datetime.py
